add authorization, connect survey to user table

+++ src/db/index.ts
@@ -1,7 +1,7 @@
 // import "@/lib/config";
 import { auth } from "@/lib/auth";
 import { sql } from "@vercel/postgres";
-import { eq } from "drizzle-orm";
+import { and, eq } from "drizzle-orm";
 import { drizzle } from "drizzle-orm/vercel-postgres";
 import { question as questionTable, survey as surveyTable, users } from "./schema";
 
@@ -12,6 +12,7 @@ export const db = drizzle(sql);
 export type Survey = {
   id: number;
   name: string;
+  userId: string;
   updated_at: Date | null;
   created_at: Date;
   deleted_at: Date | null;
@@ -34,7 +35,10 @@ export async function createSurvey(name: string) {
   const session = await auth();
   if (!session?.user) return { survey: null, message: "unauthenticated" };
 
-  const survey = await db.insert(surveyTable).values({ name }).returning();
+  const survey = await db
+    .insert(surveyTable)
+    .values({ name, userId: session?.user?.id! })
+    .returning();
   if (!survey || survey.length == 0) return { survey, message: "internal error" };
   return { survey, message: "success" };
 }
@@ -47,6 +51,7 @@ export async function getSurveys() {
   const surveys = await db
     .select()
     .from(surveyTable)
+    .where(eq(surveyTable.userId, session?.user?.id!))
     .leftJoin(questionTable, eq(surveyTable.id, questionTable.surveyId));
 
   // aggregate the questions for each survey
@@ -72,6 +77,10 @@ export async function getSurveyById(id: number) {
   if (!session?.user) return { survey: null, message: "unauthenticated" };
 
   const survey = await db.select().from(surveyTable).where(eq(surveyTable.id, id));
+
+  if (survey.length > 0 && survey[0].userId !== session?.user?.id) {
+    return { message: "unauthorized" };
+  }
   if (!survey || survey.length == 0) return { survey, message: "not found" };
   return { survey, message: "success" };
 }
@@ -83,7 +92,7 @@ export async function updateSurvey(id: number, name: string) {
   const survey = await db
     .update(surveyTable)
     .set({ name })
-    .where(eq(surveyTable.id, id))
+    .where(and(eq(surveyTable.id, id), eq(surveyTable.userId, session?.user?.id!)))
     .returning();
   if (!survey || survey.length == 0) return { survey, message: "not found" };
   return { survey, message: "success" };
@@ -93,7 +102,10 @@ export async function deleteSurvey(id: number) {
   const session = await auth();
   if (!session?.user) return { survey: null, message: "unauthenticated" };
 
-  const survey = await db.delete(surveyTable).where(eq(surveyTable.id, id)).returning();
+  const survey = await db
+    .delete(surveyTable)
+    .where(and(eq(surveyTable.id, id), eq(surveyTable.userId, session?.user?.id!)))
+    .returning();
   if (!survey || survey.length == 0) return { survey, message: "not found" };
   return { survey, message: "success" };
 }
@@ -103,6 +115,10 @@ export async function deleteSurvey(id: number) {
 export async function getQuestionsForSurvey(surveyId: number) {
   const session = await auth();
   if (!session?.user) return { questions: [], message: "unauthenticated" };
+
+  if (!(await isAuthorized(surveyId, session?.user))) {
+    return { question: [], message: "unauthorized" };
+  }
 
   const questions = await db
     .select()
@@ -116,15 +132,23 @@ export async function createQuestion(questionText: string, surveyId: number) {
   const session = await auth();
   if (!session?.user) return { question: null, message: "unauthenticated" };
 
+  if (!(await isAuthorized(surveyId, session?.user))) {
+    return { question: null, message: "unauthorized" };
+  }
+
   const question = await db.insert(questionTable).values({ questionText, surveyId }).returning();
   if (!question || question.length == 0) return { question, message: "internal error" };
 
   return { question, message: "success" };
 }
 
-export async function updateQuestion(id: number, questionText: string) {
+export async function updateQuestion(id: number, questionText: string, surveyId: number) {
   const session = await auth();
   if (!session?.user) return { question: null, message: "unauthenticated" };
+
+  if (!(await isAuthorized(surveyId, session?.user))) {
+    return { question: null, message: "unauthorized" };
+  }
 
   const question = await db
     .update(questionTable)
@@ -136,18 +160,26 @@ export async function updateQuestion(id: number, questionText: string) {
   return { question, message: "success" };
 }
 
-export async function deleteQuestion(id: number) {
+export async function deleteQuestion(id: number, surveyId: number) {
   const session = await auth();
   if (!session?.user) return { question: null, message: "unauthenticated" };
+
+  if (!(await isAuthorized(surveyId, session?.user))) {
+    return { question: null, message: "unauthorized" };
+  }
 
   const question = await db.delete(questionTable).where(eq(questionTable.id, id)).returning();
   if (!question || question.length == 0) return { question, message: "not found" };
   return { question, message: "success" };
 }
 
-export async function getQuestionById(id: number) {
+export async function getQuestionById(id: number, surveyId: number) {
   const session = await auth();
   if (!session?.user) return { question: null, message: "unauthenticated" };
+
+  if (!(await isAuthorized(surveyId, session?.user))) {
+    return { question: null, message: "unauthorized" };
+  }
 
   const question = await db.select().from(questionTable).where(eq(questionTable.id, id));
   if (!question || question.length == 0) return { question, message: "not found" };
@@ -160,3 +192,13 @@ export const getUsers = async () => {
   const selectResult = await db.select().from(users);
   return selectResult;
 };
+
+// HELPER FUNCTIONS
+
+async function isAuthorized(surveyId: number, user: any) {
+  const survey = await db.select().from(surveyTable).where(eq(surveyTable.id, surveyId));
+  if (survey.length > 0 && survey[0].userId !== user.id) {
+    return false;
+  }
+  return true;
+}
